Develop

+++ BL/Repositories/DeveloperRepository.cs
@@ -1,5 +1,4 @@
-﻿
-using eFolio.EF;
+﻿using eFolio.EF;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -10,13 +9,9 @@ namespace eFolio.BL
     {
         private eFolioDBContext db;
 
-        public DeveloperRepository(string connectionString)
+        public DeveloperRepository(eFolioDBContext eFolioDBContext)
         {
-            //var optionsBuilder = new DbContextOptionsBuilder<eFolioDBContext>();
-
-            //var options = optionsBuilder.UseSqlServer(connectionString).Options;
-
-            //this.db = new eFolioDBContext(options);
+            this.db = eFolioDBContext;
         }
 
         public void Add(DeveloperEntity item)
@@ -29,16 +24,12 @@ namespace eFolio.BL
         public void Delete(int id)
         {
             DeveloperEntity developer = db.Developers.Find(id);
-            db.Developers.Remove(developer);
 
-            db.SaveChanges();
-        }
-
-        public DeveloperEntity GetItem(int id)
-        {
-            DeveloperEntity developer = db.Developers.Find(id);
-
-            return developer;
+            if (developer != null)
+            {
+                db.Developers.Remove(developer);
+                db.SaveChanges();
+            }
         }
 
         public IEnumerable<DeveloperEntity> GetItemsList()
@@ -46,9 +37,11 @@ namespace eFolio.BL
             return db.Developers.ToListAsync().Result;
         }
 
-        public IEnumerable<DeveloperEntity> Search(string request)
+        public DeveloperEntity GetItem(int id)
         {
-            throw new NotImplementedException();
+            var developerEntity = db.Developers.Find(id);
+
+            return developerEntity;
         }
 
         public void Update(DeveloperEntity item)
